Introduction
I would like to share regarding the 1st Project Assignment experience at
my on-going Virtual Internship at Feynn Labs AI Incubator and Feynn Labs. During this period,
I have come up with Dynamic Pricing in Small Business using Machine Learning.
As per the AI Product/Service Ideation task, I have to come up with a business idea where Machine Learning/Data Science will be applied in small or medium businesses and help them with their sales, business operations, marketing etc. and prepare a report on the abstract prototype/business model.

The e-commerce website owners face a huge problem while setting the price of their products. And once they set it, another problem arises that it cannot be changed meanwhile it is required to change with time. Here's why, the possibility of a product being sold depends on many factors and these factors change with time. So the price of the product cannot be fixed and has to be changed dynamically.
The project here solves this problem by collecting customer data and using a Machine Learning model to dynamically set the price of a product changing from time to time.

Price is, unsurprisingly, one of the most important factors when considering a purchase. In fact, price is one of the major drivers for at least 47% of customers.
There are many other factors that may influence your prices – supply and demand, which promotions are running, what time of the day or year it is and so on.
Via machine learning, these factors could be assessed in a fraction of a second and your site will display dynamic pricing: showing the customer the most up-to-date price for them based on the aforementioned factors